[Migration] Replace SQLite with ClientPreference + fix !zammo (#14)
- fix !zammo in case player using after roundend but not start new round yet
- migrant to client preference

+++ ZombieModSharp/Core/Modules/PlayerManager.cs
@@ -72,7 +72,7 @@ public class Player
 
     // sound section.
     public bool SoundEnabled { get; set; }
-    public float SoundVolume { get; set; } = 100.0f;
+    public int SoundVolume { get; set; } = 100;
     public Guid MoanTimer { get; set; }
     public float ZombiePainTime { get; set; }
